InitialUpload
Includes:
YarnController - From a game jam, creates a "yarn like" behavior for an object with a Unity LineRenderer Component. Attach a spool and a "player" with a controller and this yarn renderer will wrap around objects.
VideoLessonManager - From a professional project, this file takes from our Referee Training Lessons and manages the display of the videos associated with those lessons.
DialogManager/DialogPoints - From a school project,  this is a dialog system I wrote for a school project where we created a visual novel. The dialog is put into a queue and then typed out in the ui char by char. This manager allowed for "non-coding" members of the project team to create dialog within the unity editor from the inspector.